Time Applet Window design implementation
