+++ cosmic-applet-time/Cargo.toml
@@ -9,3 +9,8 @@ icon-loader = { version = "0.3.6", features = ["gtk"] }
 libcosmic.workspace = true
 nix = "0.26.2"
 chrono = { version = "0.4.23", features = ["clock"] }
+once_cell = "1"
+# Application i18n
+i18n-embed = { version = "0.13.4", features = ["fluent-system", "desktop-requester"] }
+i18n-embed-fl = "0.6.4"
+rust-embed = "6.3.0"
