+++ cosmic-applet-time/i18n.toml
@@ -1,0 +1,4 @@
+fallback_language = "en"
+
+[fluent]
+assets_dir = "i18n"
