+++ cosmic-applet-time/i18n/en/cosmic_applet_time.ftl
@@ -1,0 +1,1 @@
+datetime-settings = Date, Time and Calendar Settings...
